Continue refactor of kafka/avro out of python engine
