+++ python/hsfs/engine/python.py
@@ -33,7 +33,6 @@ from pathlib import Path
 from typing import (
     TYPE_CHECKING,
     Any,
-    Callable,
     Dict,
     List,
     Literal,
@@ -46,7 +45,6 @@ from typing import (
 if TYPE_CHECKING:
     import great_expectations
 
-import avro
 import boto3
 import hsfs
 import numpy as np
@@ -55,7 +53,7 @@ import polars as pl
 import pyarrow as pa
 import pytz
 from botocore.response import StreamingBody
-from confluent_kafka import Consumer, KafkaError, Producer, TopicPartition
+from confluent_kafka import KafkaError
 from hsfs import (
     client,
     feature,
@@ -65,7 +63,6 @@ from hsfs import (
     util,
 )
 from hsfs import storage_connector as sc
-from hsfs.client import hopsworks
 from hsfs.client.exceptions import FeatureStoreException
 from hsfs.constructor import query
 from hsfs.core import (
@@ -98,15 +95,6 @@ from tqdm.auto import tqdm
 
 # Disable pyhive INFO logging
 logging.getLogger("pyhive").setLevel(logging.WARNING)
-
-HAS_FAST = False
-try:
-    from fastavro import schemaless_writer
-    from fastavro.schema import parse_schema
-
-    HAS_FAST = True
-except ImportError:
-    pass
 
 if HAS_GREAT_EXPECTATIONS:
     import great_expectations
@@ -1292,52 +1280,6 @@ class Engine:
     ) -> np.ndarray:
         return feature_dataframe[feature_name].unique()
 
-    def _init_kafka_producer(
-        self,
-        feature_group: Union[FeatureGroup, ExternalFeatureGroup],
-        offline_write_options: Dict[str, Any],
-    ) -> Producer:
-        # setup kafka producer
-        return Producer(
-            self._get_kafka_config(
-                feature_group.feature_store_id, offline_write_options
-            )
-        )
-
-    def _init_kafka_consumer(
-        self,
-        feature_group: Union[FeatureGroup, ExternalFeatureGroup],
-        offline_write_options: Dict[str, Any],
-    ) -> Consumer:
-        # setup kafka consumer
-        consumer_config = self._get_kafka_config(
-            feature_group.feature_store_id, offline_write_options
-        )
-        if "group.id" not in consumer_config:
-            consumer_config["group.id"] = "hsfs_consumer_group"
-
-        return Consumer(consumer_config)
-
-    def _init_kafka_resources(
-        self,
-        feature_group: Union[FeatureGroup, ExternalFeatureGroup],
-        offline_write_options: Dict[str, Any],
-    ) -> Tuple[Producer, Dict[str, Callable], Callable]:
-        # setup kafka producer
-        producer = self._init_kafka_producer(feature_group, offline_write_options)
-
-        # setup complex feature writers
-        feature_writers = {
-            feature: self._get_encoder_func(
-                feature_group._get_feature_avro_schema(feature)
-            )
-            for feature in feature_group.get_complex_features()
-        }
-
-        # setup row writer function
-        writer = self._get_encoder_func(feature_group._get_encoded_avro_schema())
-        return producer, feature_writers, writer
-
     def _write_dataframe_kafka(
         self,
         feature_group: Union[FeatureGroup, ExternalFeatureGroup],
@@ -1478,106 +1420,6 @@ class Engine:
         if isinstance(feature_group, ExternalFeatureGroup):
             return None
         return feature_group.materialization_job
-
-    def _kafka_get_offsets(
-        self,
-        feature_group: Union[FeatureGroup, ExternalFeatureGroup],
-        offline_write_options: Dict[str, Any],
-        high: bool,
-    ) -> str:
-        topic_name = feature_group._online_topic_name
-        consumer = self._init_kafka_consumer(feature_group, offline_write_options)
-        topics = consumer.list_topics(
-            timeout=offline_write_options.get("kafka_timeout", 6)
-        ).topics
-        if topic_name in topics.keys():
-            # topic exists
-            offsets = ""
-            tuple_value = int(high)
-            for partition_metadata in topics.get(topic_name).partitions.values():
-                partition = TopicPartition(
-                    topic=topic_name, partition=partition_metadata.id
-                )
-                offsets += f",{partition_metadata.id}:{consumer.get_watermark_offsets(partition)[tuple_value]}"
-            consumer.close()
-
-            return f" -initialCheckPointString {topic_name + offsets}"
-        return ""
-
-    def _kafka_produce(
-        self,
-        producer: Producer,
-        feature_group: Union[FeatureGroup, ExternalFeatureGroup],
-        key: str,
-        encoded_row: bytes,
-        acked: callable,
-        offline_write_options: Dict[str, Any],
-    ) -> None:
-        while True:
-            # if BufferError is thrown, we can be sure, message hasn't been send so we retry
-            try:
-                # produce
-                header = {
-                    "projectId": str(feature_group.feature_store.project_id).encode(
-                        "utf8"
-                    ),
-                    "featureGroupId": str(feature_group._id).encode("utf8"),
-                    "subjectId": str(feature_group.subject["id"]).encode("utf8"),
-                }
-
-                producer.produce(
-                    topic=feature_group._online_topic_name,
-                    key=key,
-                    value=encoded_row,
-                    callback=acked,
-                    headers=header,
-                )
-
-                # Trigger internal callbacks to empty op queue
-                producer.poll(0)
-                break
-            except BufferError as e:
-                if offline_write_options.get("debug_kafka", False):
-                    print("Caught: {}".format(e))
-                # backoff for 1 second
-                producer.poll(1)
-
-    def _encode_complex_features(
-        self, feature_writers: Dict[str, callable], row: Dict[str, Any]
-    ) -> Dict[str, Any]:
-        for feature_name, writer in feature_writers.items():
-            with BytesIO() as outf:
-                writer(row[feature_name], outf)
-                row[feature_name] = outf.getvalue()
-        return row
-
-    def _get_encoder_func(self, writer_schema: str) -> callable:
-        if HAS_FAST:
-            schema = json.loads(writer_schema)
-            parsed_schema = parse_schema(schema)
-            return lambda record, outf: schemaless_writer(outf, parsed_schema, record)
-
-        parsed_schema = avro.schema.parse(writer_schema)
-        writer = avro.io.DatumWriter(parsed_schema)
-        return lambda record, outf: writer.write(record, avro.io.BinaryEncoder(outf))
-
-    def _get_kafka_config(
-        self, feature_store_id: int, write_options: Optional[Dict[str, Any]] = None
-    ) -> Dict[str, Any]:
-        if write_options is None:
-            write_options = {}
-        external = not (
-            isinstance(client.get_instance(), hopsworks.Client)
-            or write_options.get("internal_kafka", False)
-        )
-
-        storage_connector = self._storage_connector_api.get_kafka_connector(
-            feature_store_id, external
-        )
-
-        config = storage_connector.confluent_options()
-        config.update(write_options.get("kafka_producer_config", {}))
-        return config
 
     @staticmethod
     def _convert_pandas_dtype_to_offline_type(arrow_type: str) -> str:
